refactor: 벌레 컨트롤러 및 테스트 패키지 구조 변경 (#97)
* refactor: 벌레 상품 조회 API URL 변경

* style: 테스트 패키지 구조 변경
