+++ src/main/java/com/moabam/api/application/bug/BugService.java
@@ -2,6 +2,7 @@ package com.moabam.api.application.bug;
 
 import static com.moabam.api.domain.bug.BugActionType.*;
 import static com.moabam.api.domain.bug.BugType.*;
+import static com.moabam.api.domain.product.ProductType.*;
 
 import java.util.List;
 
@@ -9,12 +10,16 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.moabam.api.application.member.MemberService;
+import com.moabam.api.application.product.ProductMapper;
 import com.moabam.api.domain.bug.BugHistory;
 import com.moabam.api.domain.bug.BugType;
 import com.moabam.api.domain.bug.repository.BugHistorySearchRepository;
 import com.moabam.api.domain.member.Member;
+import com.moabam.api.domain.product.Product;
+import com.moabam.api.domain.product.repository.ProductRepository;
 import com.moabam.api.dto.TodayBugResponse;
 import com.moabam.api.dto.bug.BugResponse;
+import com.moabam.api.dto.product.ProductsResponse;
 import com.moabam.global.common.util.ClockHolder;
 
 import lombok.RequiredArgsConstructor;
@@ -26,6 +31,7 @@ public class BugService {
 
 	private final MemberService memberService;
 	private final BugHistorySearchRepository bugHistorySearchRepository;
+	private final ProductRepository productRepository;
 	private final ClockHolder clockHolder;
 
 	public BugResponse getBug(Long memberId) {
@@ -40,6 +46,12 @@ public class BugService {
 		int nightBug = calculateBugQuantity(todayRewardBug, NIGHT);
 
 		return BugMapper.toTodayBugResponse(morningBug, nightBug);
+	}
+
+	public ProductsResponse getBugProducts() {
+		List<Product> products = productRepository.findAllByType(BUG);
+
+		return ProductMapper.toProductsResponse(products);
 	}
 
 	private int calculateBugQuantity(List<BugHistory> bugHistory, BugType bugType) {
