+++ src/main/java/com/moabam/api/presentation/BugController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.moabam.api.application.bug.BugService;
 import com.moabam.api.dto.TodayBugResponse;
 import com.moabam.api.dto.bug.BugResponse;
+import com.moabam.api.dto.product.ProductsResponse;
 import com.moabam.global.auth.annotation.CurrentMember;
 import com.moabam.global.auth.model.AuthorizationMember;
 
@@ -31,5 +32,11 @@ public class BugController {
 	@ResponseStatus(HttpStatus.OK)
 	public TodayBugResponse getTodayBug(@CurrentMember AuthorizationMember member) {
 		return bugService.getTodayBug(member.id());
+	}
+
+	@GetMapping("/products")
+	@ResponseStatus(HttpStatus.OK)
+	public ProductsResponse getBugProducts() {
+		return bugService.getBugProducts();
 	}
 }
