--- src/main/java/com/moabam/api/presentation/ProductController.java
@@ -1,26 +1,0 @@
-package com.moabam.api.presentation;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.moabam.api.application.product.ProductService;
-import com.moabam.api.dto.product.ProductsResponse;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequestMapping("/products")
-@RequiredArgsConstructor
-public class ProductController {
-
-	private final ProductService productService;
-
-	@GetMapping
-	@ResponseStatus(HttpStatus.OK)
-	public ProductsResponse getProducts() {
-		return productService.getProducts();
-	}
-}
