+++ src/test/java/com/moabam/api/domain/bug/BugTest.java
@@ -1,4 +1,4 @@
-package com.moabam.api.domain.entity;
+package com.moabam.api.domain.bug;
 
 import static com.moabam.support.fixture.BugFixture.*;
 import static org.assertj.core.api.Assertions.*;
@@ -10,8 +10,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import com.moabam.api.domain.bug.Bug;
-import com.moabam.api.domain.bug.BugType;
 import com.moabam.global.error.exception.BadRequestException;
 
 class BugTest {
