+++ src/test/java/com/moabam/api/domain/item/ItemTest.java
@@ -1,4 +1,4 @@
-package com.moabam.api.domain.entity;
+package com.moabam.api.domain.item;
 
 import static com.moabam.support.fixture.ItemFixture.*;
 import static org.assertj.core.api.Assertions.*;
@@ -11,7 +11,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import com.moabam.api.domain.bug.BugType;
-import com.moabam.api.domain.item.Item;
 import com.moabam.global.error.exception.BadRequestException;
 
 class ItemTest {
