--- src/test/java/com/moabam/api/presentation/ProductControllerTest.java
@@ -1,60 +1,0 @@
-package com.moabam.api.presentation;
-
-import static com.moabam.support.fixture.ProductFixture.*;
-import static java.nio.charset.StandardCharsets.*;
-import static org.assertj.core.api.Assertions.*;
-import static org.springframework.http.MediaType.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-
-import java.util.List;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.moabam.api.application.product.ProductMapper;
-import com.moabam.api.domain.product.Product;
-import com.moabam.api.domain.product.repository.ProductRepository;
-import com.moabam.api.dto.product.ProductsResponse;
-import com.moabam.support.common.WithoutFilterSupporter;
-
-@Transactional
-@SpringBootTest
-@AutoConfigureMockMvc
-class ProductControllerTest extends WithoutFilterSupporter {
-
-	@Autowired
-	MockMvc mockMvc;
-
-	@Autowired
-	ObjectMapper objectMapper;
-
-	@Autowired
-	ProductRepository productRepository;
-
-	@DisplayName("상품 목록을 조회한다.")
-	@Test
-	void get_products_success() throws Exception {
-		// given
-		List<Product> products = productRepository.saveAll(List.of(bugProduct(), bugProduct()));
-		ProductsResponse expected = ProductMapper.toProductsResponse(products);
-
-		// when, then
-		String content = mockMvc.perform(get("/products")
-				.contentType(APPLICATION_JSON))
-			.andExpect(status().isOk())
-			.andDo(print())
-			.andReturn()
-			.getResponse()
-			.getContentAsString(UTF_8);
-		ProductsResponse actual = objectMapper.readValue(content, ProductsResponse.class);
-		assertThat(actual).isEqualTo(expected);
-	}
-}
